API/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
werkzeug==2.3.7
sqlalchemy==2.0.21
